requirements.txt
asgiref==3.3.1
Django==3.0.11
django-annoying==0.10.6
django-cors-headers==3.5.0
django-random-queryset==0.1.3
djangorestframework==3.12.2
gunicorn==20.1.0
numpy==1.19.4
pandas==1.1.4
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
xlrd==2.0.1
